refactor(missions): extract shared MissionGoalRow into a neutral module
Code-review follow-up: MissionCard (a non-popup board component) was
reaching into popups/MissionPopup for MissionGoalRow + the goal/reward
view-model types, and missionView.ts split its type imports awkwardly
across the popup. Move MissionGoalRow, MissionGoalVM and MissionRewardVM
into scripts/components/MissionGoal.tsx so both the dialog and the card
depend on a neutral leaf module — no card→popup dependency, no cycle.
Pure relocation: rendered DOM is unchanged.
